# Multiselect perf: send updates only to loaded items